+++ ml_api/Dockerfile
@@ -12,8 +12,4 @@ COPY ./src /code/src
 COPY ./app /code/app
 COPY ./models /code/models
 
-COPY .env /code/.env
-
-EXPOSE 8000
-
 CMD ["uvicorn", "app.main:app", "--forwarded-allow-ips", "*", "--host", "0.0.0.0", "--port", "80"]
